Refactor codebase and fix linter errors
